fix(build): resolve gcc -Werror sign/conversion failures in audio-io and routing
GCC 13 (ubuntu-latest default) rejects the implicit size_t -> int64_t
accumulations in waveform_processor.cpp and the int -> uint8_t output-bus
arithmetic in routing_matrix.cpp under -Werror=sign-conversion/-Werror=conversion.
Add the explicit casts so Debug builds compile clean with GCC.

+++ src/core/routing/routing_matrix.cpp
@@ -636,8 +636,8 @@ SessionGraphError RoutingMatrix::processRouting(const float* const* channel_inpu
     // output_bus 1 = channels 2-3 (stereo pair 2)
     // etc.
     uint8_t output_bus = group.config.output_bus;
-    uint8_t out_L = output_bus * 2;
-    uint8_t out_R = output_bus * 2 + 1;
+    uint8_t out_L = static_cast<uint8_t>(output_bus * 2);
+    uint8_t out_R = static_cast<uint8_t>(output_bus * 2 + 1);
 
     // Validate output channels exist
     if (out_R >= config.num_outputs) {
